requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
greenlet==3.0.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.26.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.17.0
polars==0.19.3
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.2
tenacity==8.2.3
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
